atoi modified for int max

+++ libft/ft_atoi.c
@@ -11,9 +11,9 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
-	int	nbr;
+	long int	nbr;
 	int	sign;
 	int	i;
